style: remove unnecessary u64 casts in disk space check
The f_bavail and f_frsize fields are already u64 on Linux, making
the explicit casts redundant. Fixes clippy unnecessary_cast warnings.

+++ crates/cli/src/cli/doctor.rs
@@ -653,7 +653,7 @@ fn check_disk_space() -> CheckResult {
             unsafe {
                 let mut stat: nix::libc::statvfs = std::mem::zeroed();
                 if nix::libc::statvfs(path.as_ptr(), &mut stat) == 0 {
-                    let free_bytes = (stat.f_bavail as u64) * (stat.f_frsize as u64);
+                    let free_bytes = stat.f_bavail * stat.f_frsize;
                     let free_mb = free_bytes / (1024 * 1024);
 
                     if free_mb < 100 {
